fix test compilation
