+++ test/routes/forms/submission/router.test.ts
@@ -4,6 +4,7 @@ import express, { type Express } from "express";
 import { submissionApiRoute } from "@routes/forms/submission/router";
 import { getFormSubmission } from "@src/lib/vault/getFormSubmission";
 import { FormSubmissionStatus } from "@src/lib/vault/dataStructures/formSubmission";
+import { buildMockedFormSubmission } from "test/mocks/formSubmission";
 
 vi.mock("@lib/vault/getFormSubmission");
 const getFormSubmissionMock = vi.mocked(getFormSubmission);
@@ -26,10 +27,9 @@ describe("/forms/:formId/submission", () => {
   describe("/:submissionName", () => {
     describe("Response to GET operation when", () => {
       it("submissionName format is valid", async () => {
-        getFormSubmissionMock.mockResolvedValueOnce({
-          status: FormSubmissionStatus.New,
-          answers: "Here is my form submission",
-        });
+        getFormSubmissionMock.mockResolvedValueOnce(
+          buildMockedFormSubmission(FormSubmissionStatus.New),
+        );
 
         const response = await request(server).get("/01-08-a571");
