+++ test/routes/forms/submission/submissionName/router.test.ts
@@ -3,23 +3,11 @@ import request from "supertest";
 import express, { type Express } from "express";
 import { submissionNameApiRoute } from "@routes/forms/submission/submissionName/router";
 import { getFormSubmission } from "@lib/vault/getFormSubmission";
-import {
-  type FormSubmission,
-  FormSubmissionStatus,
-} from "@src/lib/vault/dataStructures/formSubmission";
+import { FormSubmissionStatus } from "@src/lib/vault/dataStructures/formSubmission";
+import { buildMockedFormSubmission } from "test/mocks/formSubmission";
 
 vi.mock("@lib/vault/getFormSubmission");
 const getFormSubmissionMock = vi.mocked(getFormSubmission);
-
-function buildMockedFormSubmission(
-  status: FormSubmissionStatus = FormSubmissionStatus.New,
-): FormSubmission {
-  return {
-    status,
-    confirmationCode: "58386068-6ce8-4e4f-89b2-e329df9c8b42",
-    answers: "Here is my form submission",
-  };
-}
 
 describe("/forms/:formId/submission/:submissionName", () => {
   let server: Express;
